Reorganized code to be clearer. Renamed synchronous name space to compatibility_synchronous
