+++ src/main/java/com/instructure/canvasapi/api/AnnouncementAPI.java
@@ -29,12 +29,18 @@ public class AnnouncementAPI {
         @GET("/{next}")
         void getNextPageAnnouncementsList(@EncodedPath("next") String nextURL, Callback<DiscussionTopicHeader[]> callback);
     }
-
+    /////////////////////////////////////////////////////////////////////////
+    // Build Interface Helpers
+    /////////////////////////////////////////////////////////////////////////
 
     private static AnnouncementsInterface buildInterface(CanvasCallback<?> callback, CanvasContext canvasContext) {
         RestAdapter restAdapter = CanvasRestAdapter.buildAdapter(callback, canvasContext);
         return restAdapter.create(AnnouncementsInterface.class);
     }
+
+    /////////////////////////////////////////////////////////////////////////
+    // API Calls
+    /////////////////////////////////////////////////////////////////////////
 
     public static void getFirstPageAnnouncements(CanvasContext canvasContext, CanvasCallback<DiscussionTopicHeader[]> callback) {
         if (APIHelpers.paramIsNull(callback, canvasContext)) { return; }
