+++ src/main/java/com/instructure/canvasapi/api/ConversationAPI.java
@@ -45,18 +45,12 @@ public class ConversationAPI {
         @GET("/conversations/?interleave_submissions=1")
         void getFirstPageConversationList(@Query("scope") String scope, Callback<Conversation[]> callback);
 
-        @GET("/conversations/?interleave_submissions=1")
-        Conversation[] getFirstPageConversationList(@Query("scope") String scope, @Query("per_page") int number);
 
         @GET("/{next}")
         void getNextPageConversationList(@EncodedPath("next") String nextURL, Callback<Conversation[]>callback);
 
         @GET("/conversations/{id}/?interleave_submissions=1")
         void getDetailedConversation(@Path("id") long conversation_id, @Query("auto_mark_as_read") int markAsRead, Callback<Conversation> callback);
-
-        @GET("/conversations/{id}/?interleave_submissions=1")
-        Conversation getDetailedConversationSynchronous(@Path("id") long conversation_id);
-
 
         @POST("/conversations/{id}/add_message")
         void addMessageToConversation(@Path("id")long conversation_id, @Query("body")String message, CanvasCallback<Conversation> callback);
@@ -75,7 +69,22 @@ public class ConversationAPI {
 
         @PUT("/conversations/{conversationid}?conversation[workflow_state]=read")
         void unArchiveConversation(@Path("conversationid")long conversationID, CanvasCallback<Response>responseCallback);
+
+        /////////////////////////////////////////////////////////////////////////////
+        // Synchronous
+        /////////////////////////////////////////////////////////////////////////////
+
+        @GET("/conversations/?interleave_submissions=1")
+        Conversation[] getFirstPageConversationList(@Query("scope") String scope, @Query("per_page") int number);
+
+        @GET("/conversations/{id}/?interleave_submissions=1")
+        Conversation getDetailedConversationSynchronous(@Path("id") long conversation_id);
+
     }
+
+    /////////////////////////////////////////////////////////////////////////
+    // Build Interface Helpers
+    /////////////////////////////////////////////////////////////////////////
 
     private static ConversationsInterface buildInterface(CanvasCallback<?> callback) {
         return buildInterface(callback.getContext());
@@ -86,21 +95,15 @@ public class ConversationAPI {
         return restAdapter.create(ConversationsInterface.class);
     }
 
+    /////////////////////////////////////////////////////////////////////////
+    // API Calls
+    /////////////////////////////////////////////////////////////////////////
 
     public static void getDetailedConversation(CanvasCallback<Conversation> callback, long conversation_id, boolean markAsRead) {
         if (APIHelpers.paramIsNull(callback)) return;
 
         callback.readFromCache(getDetailedConversationCacheFilename(conversation_id));
         buildInterface(callback).getDetailedConversation(conversation_id, APIHelpers.booleanToInt(markAsRead), callback);
-    }
-    public static Conversation getDetailedConversationSynchronous(Context context, long conversation_id){
-        //If not able to parse (no network for example), this will crash. Handle that case.
-        try {
-            return buildInterface(context).getDetailedConversationSynchronous(conversation_id);
-        } catch (Exception E){
-            return null;
-        }
-
     }
 
     public static void getFirstPageConversations(CanvasCallback<Conversation[]> callback, ConversationScope scope) {
@@ -110,16 +113,6 @@ public class ConversationAPI {
         buildInterface(callback).getFirstPageConversationList(conversationScopeToString(scope), callback);
     }
 
-    public static Conversation[] getFirstPageConversationsSynchronous(ConversationScope scope, Context context, int numberToReturn) {
-
-        try{
-            return buildInterface(context).getFirstPageConversationList(conversationScopeToString(scope), numberToReturn);
-        } catch (Exception E){
-            return null;
-        }
-
-
-    }
 
     public static void getNextPageConversations(CanvasCallback<Conversation[]> callback, String nextURL){
         if (APIHelpers.paramIsNull(callback, nextURL)) return;
@@ -179,5 +172,32 @@ public class ConversationAPI {
 
         buildInterface(responseCanvasCallback).unArchiveConversation(conversationId,responseCanvasCallback);
     }
+
+    /////////////////////////////////////////////////////////////////////////////
+    // Synchronous
+    //
+    // If Retrofit is unable to parse (no network for example) Synchronous calls
+    // will throw a nullPointer exception. All synchronous calls need to be in a
+    // try catch block.
+    /////////////////////////////////////////////////////////////////////////////
+
+    public static Conversation[] getFirstPageConversationsSynchronous(ConversationScope scope, Context context, int numberToReturn) {
+
+        try{
+            return buildInterface(context).getFirstPageConversationList(conversationScopeToString(scope), numberToReturn);
+        } catch (Exception E){
+            return null;
+        }
+    }
+
+    public static Conversation getDetailedConversationSynchronous(Context context, long conversation_id){
+        try {
+            return buildInterface(context).getDetailedConversationSynchronous(conversation_id);
+        } catch (Exception E){
+            return null;
+        }
+    }
+
+
 
 }
