+++ src/main/java/com/instructure/canvasapi/api/ModuleAPI.java
@@ -42,10 +42,18 @@ public class ModuleAPI {
         void getNextPageModuleItemList(@EncodedPath("next") String nextURL, Callback<ModuleItem[]> callback);
     }
 
+    /////////////////////////////////////////////////////////////////////////
+    // Build Interface Helpers
+    /////////////////////////////////////////////////////////////////////////
+
     private static ModulesInterface buildInterface(CanvasCallback<?> callback, CanvasContext canvasContext) {
         RestAdapter restAdapter = CanvasRestAdapter.buildAdapter(callback, canvasContext);
         return restAdapter.create(ModulesInterface.class);
     }
+
+    /////////////////////////////////////////////////////////////////////////
+    // API Calls
+    /////////////////////////////////////////////////////////////////////////
 
     public static void getFirstPageModuleObjects(CanvasContext canvasContext, CanvasCallback<ModuleObject[]> callback) {
         if (APIHelpers.paramIsNull(callback, canvasContext)) { return; }
