+++ src/main/java/com/instructure/canvasapi/api/QuizAPI.java
@@ -41,11 +41,18 @@ public class QuizAPI {
         void getDetailedQuizFromURL(@EncodedPath("next") String quizURL, Callback<Quiz> callback);
     }
 
+    /////////////////////////////////////////////////////////////////////////
+    // Build Interface Helpers
+    /////////////////////////////////////////////////////////////////////////
 
     private static QuizzesInterface buildInterface(CanvasCallback<?> callback, CanvasContext canvasContext) {
         RestAdapter restAdapter = CanvasRestAdapter.buildAdapter(callback, canvasContext);
         return restAdapter.create(QuizzesInterface.class);
     }
+
+    /////////////////////////////////////////////////////////////////////////
+    // API Calls
+    /////////////////////////////////////////////////////////////////////////
 
     public static void getFirstPageQuizzes(CanvasContext canvasContext, CanvasCallback<Quiz[]> callback) {
         if (APIHelpers.paramIsNull(callback, canvasContext)) { return; }
