+++ src/main/java/com/instructure/canvasapi/api/compatibility_synchronous/UploadFileSynchronousAPI.java
@@ -1,4 +1,4 @@
-package com.instructure.canvasapi.api.synchronous;
+package com.instructure.canvasapi.api.compatibility_synchronous;
 
 import android.app.Activity;
 import android.content.Context;
